# services/app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Generación de PDFs",
    "version": "1.0.0",
    "description": "API para el sistema de diseño y generación de documentos PDF"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/variables": {
      "get": {
        "tags": ["Variables"],
        "summary": "Obtener variables",
        "description": "Obtiene todas las variables, opcionalmente filtradas por plantilla",
        "parameters": [
          {
            "name": "template_id",
            "in": "query",
            "description": "ID de la plantilla para filtrar",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de variables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Variable"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Variables"],
        "summary": "Crear variable",
        "description": "Crea una nueva variable",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariableCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Variable creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Variable"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/variables/{variable_id}": {
      "parameters": [
        {
          "name": "variable_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "tags": ["Variables"],
        "summary": "Obtener variable por ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Variable encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariableResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Variables"],
        "summary": "Actualizar variable",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariableUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variable actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariableResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Variables"],
        "summary": "Eliminar variable",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Variable eliminada exitosamente"
          }
        }
      }
    },
    "/templates/{template_id}/variables/bulk": {
      "post": {
        "tags": ["Variables"],
        "summary": "Crear múltiples variables",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariableBulkCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Variables creadas exitosamente"
          }
        }
      }
    },
    "/variables/predefined": {
      "get": {
        "tags": ["Variables"],
        "summary": "Obtener variables predefinidas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de variables predefinidas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredefinedVariable"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationRule": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["length", "range", "regex", "enum", "email", "rut", "date", "currency"]
          },
          "params": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["type", "params", "message"]
      },
      "Variable": {
        "type": "object",
        "properties": {
          "variable_id": {
            "type": "integer"
          },
          "template_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "default_value": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "variable_type": {
            "type": "string",
            "enum": ["text", "number", "date", "email", "rut", "currency", "enum"]
          },
          "validation_rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationRule"
            }
          }
        }
      },
      "VariableCreate": {
        "type": "object",
        "properties": {
          "template_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "pattern": "^[a-zA-Z][a-zA-Z0-9_]*$"
          },
          "default_value": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "variable_type": {
            "type": "string"
          },
          "validation_rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationRule"
            }
          }
        },
        "required": ["template_id", "name", "variable_type"]
      },
      "VariableUpdate": {
        "type": "object",
        "properties": {
          "default_value": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "validation_rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationRule"
            }
          }
        }
      },
      "VariableResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Variable"
          },
          {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "VariableBulkCreate": {
        "type": "object",
        "properties": {
          "variables": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariableCreate"
            }
          }
        },
        "required": ["variables"]
      },
      "PredefinedVariable": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "variable_type": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "example": {
            "type": "string"
          }
        },
        "required": ["name", "description", "variable_type", "example"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}